requirements.txt
# Dependências para todos os frameworks de API

# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Flask
flask==3.0.0

# Django REST Framework
django==4.2.7
djangorestframework==3.14.0

# Sanic
sanic==23.6.0

# Tornado
tornado==6.3.3

# Falcon
falcon==3.1.1

# Bottle
bottle==0.12.25
